// app/books/models/local/books.local-model.ts
export type initialsSearchBooksType = {
  search: string;
};

export type BooksToLocalModels = {
  id: string;
  title: string;
  author: string;
  description: string;
  price: number;
  image: string;
  category: string;
  rating: number;
  isbn: string;
  publishedDate: string;
  pages: number;
};